requirements.txt
django
djangorestframework
dj-rest-auth
django-allauth
numpy
yasa
psycopg2-binary